+++ jgclark.EventHelpers/src/index.js
@@ -3,13 +3,14 @@
 //-----------------------------------------------------------------------------
 // Event Helpers
 // Jonathan Clark
-// last updated 22.7.2022, for v0.16.6
+// last updated 29.9.2023, for v0.20.4
 //-----------------------------------------------------------------------------
 
 // allow changes in plugin.json to trigger recompilation
 import pluginJson from '../plugin.json'
-import { log, logError } from '@helpers/dev'
+import { JSP, logDebug, logError } from '@helpers/dev'
 import { updateSettingData } from '@helpers/NPConfiguration'
+import { editSettings } from '@helpers/NPSettings'
 import { showMessage } from '@helpers/userInput'
 
 export { timeBlocksToCalendar } from './timeblocks'
@@ -30,14 +31,28 @@ export async function onSettingsUpdated(): Promise<void> {
 // refactor previous variables to new types
 export async function onUpdateOrInstall(): Promise<void> {
   try {
-    log(pluginJson, `${pluginID}: onUpdateOrInstall running`)
+    logDebug(pluginJson, `${pluginID}: onUpdateOrInstall running`)
     const updateSettings = updateSettingData(pluginJson)
-    log(pluginJson, `${pluginID}: onUpdateOrInstall updateSettingData code: ${updateSettings}`)
+    logDebug(pluginJson, `${pluginID}: onUpdateOrInstall updateSettingData code: ${updateSettings}`)
     if (pluginJson['plugin.lastUpdateInfo'] !== undefined) {
       await showMessage(pluginJson['plugin.lastUpdateInfo'], 'OK, thanks', `Plugin ${pluginJson['plugin.name']} updated to v${pluginJson['plugin.version']}`)
     }
   } catch (error) {
     logError(pluginJson, error)
   }
-  log(pluginJson, `${pluginID}: onUpdateOrInstall finished`)
+  logDebug(pluginJson, `${pluginID}: onUpdateOrInstall finished`)
+}
+
+/**
+ * Update Settings/Preferences (for iOS etc)
+ * Plugin entrypoint for command: "/<plugin>: Update Plugin Settings/Preferences"
+ * @author @dwertheimer
+ */
+export async function updateSettings() {
+  try {
+    logDebug(pluginJson, `updateSettings running`)
+    await editSettings(pluginJson)
+  } catch (error) {
+    logError(pluginJson, JSP(error))
+  }
 }